ab_demo/requirements.txt
aws-cdk-lib==2.60.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-synthetics-alpha
jsii>=1.14.1
cattrs>=1.1.1
cdk-nag
cfn-lint
bandit
checkov